12.3. The Values API
The stimulus_controller() Method
I love the values API and it has one sweet trick up its sleeve that we'll learn about in the next video. But the syntax, woof.
So far, we've created all of our data-controller elements by hand... because... it's pretty simple to write data-controller="color-square".
But WebpackEncoreBundle gives us a shortcut method. Check it out: inside the element where we want the data-controller to be, add {{ stimulus_controller() }} and pass the name of the controller: color-square.
Then, for the second optional argument, we can pass an array of values. We have one: color-id set to the long addToCartForm.vars line.
Celebrate by deleting data-controller and the value.
This will give us the exact same result as before. We can see it: inspect element, find data-controller and... refresh! Sweet! It does look the same as before and we still get the log.
I love this feature because the values API is incredibly powerful and this removes all the pain of using it.

Values are HTML Escaped Automatically
Oh, and this function also automatically escapes each value so it's safe to use in an HTML attribute. So if a value contains a double-quote, it won't break your page. We'll see this later when we use the values API to pass props to a React app.