13.1. The Values API
Now that we've created a colorId value, we can pass that from the server into Stimulus and read it as this.colorIdValue. Let's use that to select the color square on load.

Setting the Initial Color
Thanks to our organization, this will be no problemo. Replace the log with: if this.colorIdValue: just in case we want to make this value optional. Inside, call this.setSelectedColor() and pass this.colorIdValue.
I do love that we created that re-usable setSelectedColor() method! Let's try this: fly over to the browser and... it... doesn't work? In the console, we have a giant error:

  Cannot read property classList of undefined

... coming from setSelectedColor().
Let's go take a look. I think it's coming from right here. For some reason, the findSelectedColorSquare() method is not finding the element... which is odd.
Scroll down to it. Ahh. The problem is the stronger type that the values API gave us. element.dataset.colorId, which just uses the normal dataset functionality, will be a String. But this.selectedColorId will now be a Number... because if we scroll up, we set it to this.colorIdValue, which we know is a true Number type.