13.3. The Values API
Value Change Callback
There's one last feature about the values API that we haven't talked about yet. And it's really going to help us. In fact, it's going to let us to delete a lot of code from our controller. It's called a change callback.
Very simply, we can tell Stimulus to automatically call a function whenever a value changes, like when our colorId value changes. How? With a specially named method.
Add a new method called colorIdValueChanged(). Inside, go steal the code from earlier: this.setSelectedColor(this.colorIdValue).
And now we can remove the code inside connect().
Here's the flow: on load, the colorId value will be read from our data attribute. That will change the colorId value and cause our callback to be executed. The naming of the method is important: it must be exactly named like this for Stimulus to recognize it as a change callback.
Let's give it a go! Refresh and... yea, it did work! To make it more obvious, click red, then reload. Back to green!

Callbacks even Listen to Attribute Changes
Ready to have your mind blown? Find the data-controller element in your inspector. 2 is the id of the green item. Let's change it to 1, which is red. Woh! The selected color square changed! Our callback is even executed when the value's data-attribute is updated! That's bonkers.