13.4. The Values API
Using Values Instead of a Property
Look back at our controller. Now, I'm wondering something: do we really need both a selectedColorId property and a colorId value? Don't they both... kinda store the currently selected color?
Yep! And the answer is that we do not need both.
A value is basically a property with superpowers. Values have the ability to read an initial value from a data attribute, support change callbacks and wear a bright red cape.
Check this out: in the colorIdValueChanged() method, let's add all the logic that we need to get this to work on its own. In other words, I want to replace this.setSelectedColor() with code that does the same thing.
Start by setting the value on the select: this.selectTarget.value = this.colorIdValue.
The only other thing that we need to do inside here is a loop over the color squares to set the selected class correctly. Do that with this.colorSquareTargets.forEach() and pass this an arrow function with an element argument. Inside, we can use an if statement to figure out if we should be adding the class or removing it: if element.dataset.colorId == this.colorIdValue then we know this element is now the current color. Add the class with element.classList.add('selected'). Else, remove the selected class.
Nice! Up in selectColor(), we don't need to call setSelectedColor() anymore. Instead, just set the value! Copy the event.currentTarget code and say this.colorIdValue = and paste.
That's it! When we click a color square, selectColor will be called. Then we set this.colorIdValue and that triggers our colorIdValueChanged() method. Booya!
Test drive time! When we refresh... the initial color is selected. And when we click... that works too! We did lose the ability to click a second time to unselect a color - but we'll fix that in a minute.