13.5. The Values API
Before we do, let's celebrate by removing a ton of code! We don't need setSelectedColor() anymore... or findSelectedColorSquare()... or the selectedColorId property. I'll remove that in a minute.
If you want to get back the ability to click again to unselect a color, we can do that with a little extra logic in selectColor. Add const clickedColor equals the event.currentTarget code.
For the next line, use the ternary syntax: this.colorIdValue = , if clickedColor == this.colorIdValue - so, if the clicked color is already selected - then set it to null. Otherwise set it to clickedColor.
Test it out: refresh... then click green again. Gone!

Our Tiny Controller
Go back to your editor. This is the final version of our controller. Oh, after I remove the unused selectedColorId property... now this is the final version of our controller.
And look at it! It's less than 30 lines of code and is incredibly readable. This is how I want my JavaScript to look.
Head back to the homepage of our site. This has a functional search... but it's so boring. Next: let's add an AJAX-powered "quick search" that shows matching results under the search box as we type.