14.1. Search Suggestions
One thing we haven't done yet in stimulus is handle Ajax. Let's do that by "enhancing" the search on this page. The search here does work: it's a completely normal form that submits via a GET request. In the controller - src/Controller/ProductController.php - we read the q query parameter and use it to filter the product list.
Now, our UX team wants to make it fancier: as the user types we will render a "quick results" or "suggestions" list below the search box.

Bootstrapping the Controller & Markup
Let's get to work. Start with the Stimulus controller. So in assets/controllers/ create a new file called, how about, search-preview_controller.js.
This starts exactly like every other Stimulus controller ever: import { Controller } from 'stimulus' and export default class extends Controller.
I also like to have a connect() method so we can console.log() a message to make sure everything is... um... connected.
The template for the homepage lives at templates/product/index.html.twig. Scroll down a bit... perfect. Here's the search form.
So let's think: we could add the data-controller attribute directly to the <input>. After all, we need to do something when that input changes. But we're also going to need a place to put the new search suggestions HTML.
So let's add the controller on the <div> around the input. Break this onto multiple lines and then add {{ stimulus_controller() }} passing the name of our controller: search-preview.
Let's check it! Find your browser, refresh and check the console. Connected!