14.2. Search Suggestions
Taking Action as the User Types
Ok: each time the user types a character into the box, we need to make an Ajax request to get a list of suggested, matching products. That means we need to add an action to the input.
Do that down here with our actions syntax: data-action="", the name of our controller - search-preview - a pound sign and then the method to call. How about: onSearchInput.
I'm calling it onSearchInput because the default action for an input element is actually called... input! That's a native DOM event that happens whenever the value of this box changes. Basically, each time the user types a character.
Copy the method name then head into the controller. Let's replace the connect() method with onSearchInput(). Give it an event argument and console.log() that.
Let's check it! Go refresh and... when we type... very nice!