14.3. Search Suggestions
Generating URLs in Symfony & Passing to Stimulus with Values
Look back at the controller: src/Controller/ProductController.php, at the index() method.
This action is responsible for rendering the homepage that we're currently looking at. And it also contains the logic to filter the products based on the q search query parameter. That's because, if you look in the template, our form does not have an action attribute. This means that it will submit right back to the current URL. In other words, this submits back to the homepage, but now with the ?q= part.
So here's the plan: we're going to send the Ajax request from our Stimulus controller also to the homepage route and controller... because it already has all the search logic we need. We're also going to add a second preview query parameter. We'll use that here in a few minutes to decide if we should render the full HTML page or just the HTML for the search preview.
The point is: we're going to make the Ajax request to the homepage route. So, question: in our Stimulus controller, should we just... hard-code that URL?
We could do that. But remember! Stimulus gives us a super easy way to pass values from Symfony - like a URL that we've generated - into our controller! It's the values API.
Add static values = {} an object, and create one called url. This will be a String.
Down in the method, instead of logging the event, let's console.log(this.urlValue).
In the template, to pass this in, we can use the 2nd argument of stimulus_controller. Pass url - that's the name of the value - and set it to the normal way that we generate routes in Twig: the path() function and the name of the route, which is... let me check... app_homepage. Paste that.
Let's check it! Move over and refresh. The first thing you can see, if you inspect element, is that we have the new data-search-preview-url-value="/" attribute. And as we type... nice! It correctly logs that value.
Next: let's make the Ajax call! But... what should the Ajax endpoint return? JSON? No way! Definitely HTML... but not a full page of HTML.