17.2. Stimulus Behaviors: stimulus-use
Let's go look at the documentation for useClickOutside. I'll scroll down to "usage".
Ok: step 1 is to activate the behavior in our connect() method. Cool. Copy this line... and let's make sure the library finished downloading. It did.
Over in the controller, go to the top to import the behavior: import {} and then the behavior we need - useClickOutside.
Sweet! PhpStorm auto-completed the rest of that line for me.
Below, add a connect() method and paste: useClickOutside(this).
For step 2, look at the docs: we need to add a clickOutside() method. Ok! Let's add it at the bottom: clickOutside(event). When the user clicks outside of our controller element, we will set this.resultTarget.innerHTML = ''.
Done. Let's test it! Head back to the browser and refresh. Type a little to get some suggestions, then click off. Beautiful! And if I type again... it's back, then click off... and gone again.
People: that was like four lines of code!