17.4. Stimulus Behaviors: stimulus-use -- Debounce
But that's no problem! We just need to organize things a bit better. Try this: close up onSearchInput early and move most of the logic into a new method called async search() with a query argument.
Again, we're making this async because we have an await inside.
For onSearchInput, we don't need the async anymore... and we can now call this.search() and pass it event.currentTarget.value.
Below, set the q value to query.
This is good: we've refactored our code to have a nice, reusable search() method. And now we can change the debounce from onSearchInput to search.
Testing time! Refresh and... type real fast. Yes! Only one Ajax call.

Alright! This feature is done! Next, on the checkout page, let's add a confirmation modal when the user removes an item from the cart. For this, we'll leverage a great third party library from inside our controller: SweetAlert.