18.1. Form Submit Confirmation Controller
Let's add a few items to our cart, like a floppy disk - gotta have those - and maybe also some CD's so I can burn a mixtape for Leanna. Now head to the cart page.
A user can already remove an item from their cart. Open up the template to see how: templates/cart/cart.html.twig. Scroll down a bit... here it is... around line 50. The "remove" button is inside a form. When the user clicks, the form submits and the controller removes the item from the cart. It's super smooth and super boring. I love it!

A Massively Re-usable "Submit Confirm" Controller
But now, I need to enhance this. When the user clicks "remove", I want to open a modal where the user can confirm that they want to remove the item. In fact, this is going to be even cooler than it sounds because the Stimulus controller we're about to create will be re-usable across any form in our entire app. Want to pop up a confirmation before the user submits a checkout form... or change password form? This one controller will be able to handle all of those cases.
Let's get to work. Start by creating the Stimulus controller. In assets/controllers/, add a new file called, how about, submit-confirm_controller.js. I'm calling this submit-confirm... and not "delete-confirm" or "cart-remove-confirm" because it will be reusable on any form.
Start the normal way: import { Controller } from 'stimulus' and then export default class extends Controller with a connect() method to make sure everything is hooked up: console.log()... a dinosaur (🦖).
Next up, go activate this in the template. Adding it to the form tag should be fine: {{ stimulus_controller('submit-confirm') }}.
Let's make sure it's connected! I'll re-open my console.. refresh and... roar! We even see two dinosaurs because there are two different controllers on this page. My 4 year old son would be thrilled.