18.2. Form Submit Confirmation Controller -- yarn add sweetalert2 --dev
Hello SweetAlert2
To create the actual modal, search for sweetalert2. I love this library. It's an easy - but highly customizable - alert system. If you scroll down a bit... one of these examples is for a modal that confirms deleting something. Here it is. This is almost exactly what we want.

Let's go get this library installed. Spin over to your terminal and run:

  yarn add sweetalert2 --dev