18.5. Form Submit Confirmation Controller
Look back at the code. Here's how this works: when you click a button, the .then() callback is executed. That's why we saw that second message: on confirm, it called Swal again.
To make this actually submit, replace the Swal.fire() with this.element - which will be the form - .submit().
That's it! Oh, and if you're thinking:

  Hey! Won't this cause an infinite loop... where we call submit() and that causes a submit event... that triggers our submit action... which will then open SweetAlert again?
Fortunately... that will not happen. When you call .submit() on a form element, the form does submit, but the submit event is not dispatched. And so, our action method will not be called again. That's just how JavaScript and the DOM work - not a Stimulus thing. I say that a lot.
Anyways, let's see if this works! Refresh, click remove and this time confirm. Woohoo! The form submitted, the page reloaded, and the item is gone!
But I think we can make this even more awesome. How? By making our controller configurable - like the text that it displays - so we can truly reuse it anywhere in our app. Even in this situation, saying "yes, delete it" on the button... when you're actually removing an item from a cart... it doesn't really make sense.
And as an extra bonus, we're going to add an option to make the form submit via Ajax. That's all next.