19.5. Making a Configurable, Reusable Controller
Making the Ajax Form Submit Configurable
But I don't want this Ajax submit to always happen on all the forms where I use this confirm submit controller... because it requires extra work to, sort of, "reset" the page after the Ajax call finishes. So let's make this behavior configurable.
Over in the controller, up on values, add one more called submitAsync which will be a Boolean.
Down in submitForm(), use that: if not this.submitAsyncValue, then this.element.submit() and return.
Let's make sure the Ajax call is gone. Actually... let me add a few more items to my cart... because it's getting kind of empty. Add the sofa in all three colors... then go back to the cart. Let's remove this one and... beautiful. It's back to the full page refresh.
Now let's reactivate the Ajax submit on just this form by passing in the submitAsync value. In the template, set submitAsync to true.
At this point, we have a clean submit confirm controller that can be reused on any form. As a bonus, you can even tell it to submit the form via Ajax.
But when we submit via Ajax, we need to somehow remove the row that was just deleted. To do that, we're going to create a second controller around the entire cart area and make the two controllers communicate to each other. Teamwork? Yup, that's next.